+++ packages/sdk/CHANGELOG.md
@@ -11,7 +11,8 @@ commit that affects the code in a way that consumers might care about should inc
 
 ## [Unreleased](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.35...HEAD)
 
-No changes.
+-   Added `table.description` and `field.description`, and `description` watch key on `table` and
+    `field`.
 
 ## [0.0.35](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.34...@airtable/blocks@0.0.35) - 2019-11-11